requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
openai==1.12.0
python-dotenv==1.0.0
plotly==5.18.0
airtable==0.4.0
requests==2.32.3
aiohttp==3.9.3
asyncio==3.4.3
streamlit-aggrid==1.0.5
openpyxl==3.1.2
python-dotenv==1.0.0
pydantic==2.10.6
watchdog==3.0.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.17.1
google-cloud-bigquery-storage==2.24.0
xlsxwriter==3.2.5
gspread==6.2.1
fastapi==0.104.1
uvicorn==0.24.0
db-dtypes==1.2.0
pyarrow>=10.0.0
pytest>=8.0.0
